pyproject.toml
[project]
name = "kitelms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "bcrypt==4.0.1",
    "bson>=0.5.10",
    "fastapi[standard]>=0.116.1",
    "granian>=2.5.0",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "ruff>=0.12.8",
    "scalar-fastapi>=1.2.3",
    "sqlmodel>=0.0.24",
]

[tool.ruff]
line-length = 120
lint.ignore = ["E712"]
lint.select = ["E", "F", "W", "I"]
exclude = [
    "alembic",
    "venv",
    "frontend"
]
